+++ CHANGES.rst
@@ -53,17 +53,25 @@ Specviz2d
 API Changes
 -----------
 
+- Deprecated ``app.get_data_from_viewer`` is removed, use ``viz_helper.get_data`` instead. [#2578]
+
+- Deprecated ``app.get_subsets_from_viewer`` is removed, use ``viz_helper.get_subsets`` instead. [#2578]
+
 Cubeviz
 ^^^^^^^
 
 Imviz
 ^^^^^
 
+- Deprecated ``do_link`` argument of ``imviz.load_data`` is removed, use ``batch_load`` context manager instead. [#2578]
+
 Mosviz
 ^^^^^^
 
 Specviz
 ^^^^^^^
+
+- Deprecated ``specviz.load_spectrum`` is removed, use ``specviz.load_data`` instead. [#2578]
 
 Specviz2d
 ^^^^^^^^^